+++ jni/WiEngine/include/utils/wyUtils.h
@@ -369,7 +369,7 @@ public:
 	 * @return 经过解析后的png图片数据, RGBA8888格式, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadBMP(int resId, size_t* w, size_t* h, bool sizeOnly);
+	static char* loadBMP(int resId, float* w, float* h, bool sizeOnly);
 
 	/**
 	 * \if English
@@ -396,7 +396,7 @@ public:
 	 * @return 经过解析后的bmp图片数据, RGBA8888格式, 如果sizeOnly是true, 则返回NULL
 	 * \endif
 	 */
-	static char* loadBMP(const char* data, size_t length, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadBMP(const char* data, size_t length, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -421,7 +421,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadBMP(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadBMP(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -448,7 +448,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadBMP(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadBMP(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -469,7 +469,7 @@ public:
 	 * @return 经过解析后的png图片数据, RGBA8888格式, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadPNG(int resId, size_t* w, size_t* h, bool sizeOnly);
+	static char* loadPNG(int resId, float* w, float* h, bool sizeOnly);
 
 	/**
 	 * \if English
@@ -496,7 +496,7 @@ public:
 	 * @return 经过解析后的png图片数据, RGBA8888格式, 如果sizeOnly是true, 则返回NULL
 	 * \endif
 	 */
-	static char* loadPNG(const char* data, size_t length, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadPNG(const char* data, size_t length, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -521,7 +521,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadPNG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadPNG(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -548,7 +548,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadPNG(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadPNG(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -573,7 +573,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadJPG(FILE* f, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadJPG(FILE* f, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -600,7 +600,7 @@ public:
 	 * @return 适合于创建OpenGL贴图的数据, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadJPG(const char* path, bool isFile, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadJPG(const char* path, bool isFile, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -627,7 +627,7 @@ public:
 	 * @return 经过解析后的jpg图片数据, RGBA8888格式, 如果sizeOnly是true, 则返回NULL, 调用者需要负责释放返回的数据
 	 * \endif
 	 */
-	static char* loadJPG(const char* data, size_t length, size_t* w, size_t* h, bool sizeOnly, float scaleX, float scaleY);
+	static char* loadJPG(const char* data, size_t length, float* w, float* h, bool sizeOnly, float scaleX, float scaleY);
 
 	/**
 	 * \if English
@@ -648,7 +648,7 @@ public:
 	 * @return 经过解析后的jpg图片数据, RGBA8888格式, 如果sizeOnly是true, 则返回NULL
 	 * \endif
 	 */
-	static char* loadJPG(int resId, size_t* w, size_t* h, bool sizeOnly);
+	static char* loadJPG(int resId, float* w, float* h, bool sizeOnly);
 	
 	/**
 	 * \if English
@@ -768,7 +768,7 @@ public:
 	 * @return true表示获得大小成功, false表示失败
 	 * \endif
 	 */
-	static bool getPVRSize(const char* data, size_t length, size_t* w, size_t* h, float scale);
+	static bool getPVRSize(const char* data, size_t length, float* w, float* h, float scale);
 
 	/**
 	 * \if English
@@ -791,7 +791,7 @@ public:
 	 * @return true表示获得大小成功, false表示失败
 	 * \endif
 	 */
-	static bool getPVRSize(int resId, size_t* w, size_t* h, float* outScale);
+	static bool getPVRSize(int resId, float* w, float* h, float* outScale);
 
 	/**
 	 * \if English
@@ -812,7 +812,7 @@ public:
 	 * @return true表示获得大小成功, false表示失败
 	 * \endif
 	 */
-	static bool getPVRSize(FILE* f, size_t* w, size_t* h, float scale);
+	static bool getPVRSize(FILE* f, float* w, float* h, float scale);
 
 	/**
 	 * \if English
@@ -835,7 +835,7 @@ public:
 	 * @return true表示获得大小成功, false表示失败
 	 * \endif
 	 */
-	static bool getPVRSize(const char* path, bool isFile, size_t* w, size_t* h, float scale);
+	static bool getPVRSize(const char* path, bool isFile, float* w, float* h, float scale);
 	
 	/**
 	 * \if English
